add log add exchange update adapter

+++ WalletModule/src/main/java/com/fzm/walletmodule/vm/ExchangeViewModel.kt
@@ -5,6 +5,7 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.fzm.wallet.sdk.bean.ExchangeFee
+import com.fzm.wallet.sdk.bean.WithHold
 import com.fzm.wallet.sdk.db.entity.Coin
 import com.fzm.wallet.sdk.net.HttpResult
 import com.fzm.wallet.sdk.repo.ExchangeRepository
@@ -17,8 +18,8 @@ class ExchangeViewModel constructor(private val exchangeRepository: ExchangeRepo
     val flashExchange: LiveData<HttpResult<String>>
         get() = _flashExchange
 
-    private val _getExLimit = MutableLiveData<HttpResult<Long>>()
-    val getExLimit: LiveData<HttpResult<Long>>
+    private val _getExLimit = MutableLiveData<HttpResult<Double>>()
+    val getExLimit: LiveData<HttpResult<Double>>
         get() = _getExLimit
 
     private val _getExFee = MutableLiveData<HttpResult<ExchangeFee>>()
